requirements.txt
django==5.1.3
mysqlclient==2.2.5
asgiref==3.8.1
pytz==2024.2
sqlparse==0.5.1
djangorestframework==3.15.2
drf-yasg==1.21.8
PyJWT==2.10.1
requests==2.32.3
urllib3==2.3.0
oauthlib==3.2.2